pyproject.toml
[tool.poetry]
name = "simplesolutions-tasks"
version = "0.1.0"
description = ""
authors = ["RocketMenace <grizzlygit777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
ruff = "^0.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
